remove tree traversal widgets for lecturers
