+++ lib/presentation/screens/new_filter/widgets/lecturer_filter/widgets/lecturer_filters_list.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:silvertimetable/data/models/lecturer/lecturer_base.dart';
 import 'package:silvertimetable/logic/schedule_manager/schedule_manager_bloc.dart';
-import 'package:silvertimetable/presentation/screens/new_filter/widgets/lecturer_filter/widgets/lecturer_level_filter.dart';
+import 'package:silvertimetable/presentation/screens/new_filter/widgets/lecturer_filter/widgets/lecturer_tile.dart';
 
 // ignore: must_be_immutable
 class LecturerFiltersList extends StatefulWidget {
@@ -14,15 +15,17 @@ class LecturerFiltersList extends StatefulWidget {
 class _LecturerFiltersListState extends State<LecturerFiltersList> {
   @override
   Widget build(BuildContext context) {
-    return SingleChildScrollView(
-      child: BlocBuilder<ScheduleManagerBloc, ScheduleManagerState>(
-        builder: (context, state) {
-          return LecturerLevelFilter(
-            initiallyExpanded: true,
-            level: state.lecturersOptionsTree!,
-          );
-        },
-      ),
+    return BlocBuilder<ScheduleManagerBloc, ScheduleManagerState>(
+      builder: (context, state) {
+        final List<LecturerBase> lecturers =
+            state.schedulesIndex.values.whereType<LecturerBase>().toList();
+
+        return ListView.builder(
+          itemCount: lecturers.length,
+          itemBuilder: (context, index) =>
+              LecturerTile(lecturer: lecturers[index]),
+        );
+      },
     );
   }
 }
